+++ source/renderer/app/containers/wallet/Wallet.js
@@ -90,6 +90,7 @@ export default class Wallet extends Component<Props> {
           isActiveScreen={this.isActiveScreen}
           onWalletNavItemClick={this.handleWalletNavItemClick}
           activeItem={app.currentPage}
+          isLegacy={activeWallet.isLegacy}
           hasNotification={hasNotification}
         >
           {this.props.children}